Update of "calculateTime()" function
Impracticable condition for an alarm set to time multiple to hours.
For example, alarm set to 09:00. Current time is 08:59:59. Condition "alm_mins == mins" is checked only if "sec>59". Mins become equal to 60 right after checking "sec>59". So, the time now is 08:60:00. After this "alarm condition" is checked. It is not satisfied. Only after all, mins are set to 00.